+++ Apps/Contoso.iOS.Puppet/ModulePages/AppCenterController.designer.cs
@@ -1,84 +1,87 @@
 // WARNING
 //
-// This file has been generated automatically by Visual Studio from the outlets and
-// actions declared in your storyboard file.
-// Manual changes to this file will not be maintained.
+// This file has been generated automatically by Visual Studio to store outlets and
+// actions made in the UI designer. If it is removed, they will be lost.
+// Manual changes to this file may not be handled correctly.
 //
 using Foundation;
-using System;
 using System.CodeDom.Compiler;
 
 namespace Contoso.iOS.Puppet
 {
-    [Register ("AppCenterController")]
-    partial class AppCenterController
-    {
-        [Outlet]
-        UIKit.UILabel LogLevelLabel { get; set; }
+	[Register ("AppCenterController")]
+	partial class AppCenterController
+	{
+		[Outlet]
+		UIKit.UISwitch AppCenterEnabledSwitch { get; set; }
 
+		[Outlet]
+		UIKit.UILabel LogLevelLabel { get; set; }
 
-        [Outlet]
-        UIKit.UILabel LogWriteLevelLabel { get; set; }
+		[Outlet]
+		UIKit.UILabel LogWriteLevelLabel { get; set; }
 
+		[Outlet]
+		UIKit.UITextField LogWriteMessage { get; set; }
 
-        [Outlet]
-        UIKit.UITextField LogWriteMessage { get; set; }
+		[Outlet]
+		UIKit.UITextField LogWriteTag { get; set; }
 
+		[Outlet]
+		UIKit.UITextField StorageSizeText { get; set; }
 
-        [Outlet]
-        UIKit.UITextField LogWriteTag { get; set; }
+		[Outlet]
+		[GeneratedCode ("iOS Designer", "1.0")]
+		UIKit.UITextField UserIdTextField { get; set; }
 
+		[Action ("SaveStorageSize:")]
+		partial void SaveStorageSize (Foundation.NSObject sender);
 
-        [Outlet]
-        UIKit.UISwitch AppCenterEnabledSwitch { get; set; }
+		[Action ("UpdateEnabled")]
+		partial void UpdateEnabled ();
 
-        [Outlet]
-        [GeneratedCode ("iOS Designer", "1.0")]
-        UIKit.UITextField UserIdTextField { get; set; }
+		[Action ("UpdateUserId:")]
+		partial void UpdateUserId (UIKit.UITextField sender);
 
+		[Action ("WriteLog")]
+		partial void WriteLog ();
+		
+		void ReleaseDesignerOutlets ()
+		{
+			if (AppCenterEnabledSwitch != null) {
+				AppCenterEnabledSwitch.Dispose ();
+				AppCenterEnabledSwitch = null;
+			}
 
-        [Action ("UpdateEnabled")]
-        partial void UpdateEnabled ();
+			if (LogLevelLabel != null) {
+				LogLevelLabel.Dispose ();
+				LogLevelLabel = null;
+			}
 
+			if (LogWriteLevelLabel != null) {
+				LogWriteLevelLabel.Dispose ();
+				LogWriteLevelLabel = null;
+			}
 
-        [Action ("WriteLog")]
-        partial void WriteLog ();
+			if (LogWriteMessage != null) {
+				LogWriteMessage.Dispose ();
+				LogWriteMessage = null;
+			}
 
-        [Action ("UpdateUserId:")]
-        [GeneratedCode ("iOS Designer", "1.0")]
-        partial void UpdateUserId (UIKit.UITextField sender);
+			if (StorageSizeText != null) {
+				StorageSizeText.Dispose ();
+				StorageSizeText = null;
+			}
 
-        void ReleaseDesignerOutlets ()
-        {
-            if (AppCenterEnabledSwitch != null) {
-                AppCenterEnabledSwitch.Dispose ();
-                AppCenterEnabledSwitch = null;
-            }
+			if (LogWriteTag != null) {
+				LogWriteTag.Dispose ();
+				LogWriteTag = null;
+			}
 
-            if (LogLevelLabel != null) {
-                LogLevelLabel.Dispose ();
-                LogLevelLabel = null;
-            }
-
-            if (LogWriteLevelLabel != null) {
-                LogWriteLevelLabel.Dispose ();
-                LogWriteLevelLabel = null;
-            }
-
-            if (LogWriteMessage != null) {
-                LogWriteMessage.Dispose ();
-                LogWriteMessage = null;
-            }
-
-            if (LogWriteTag != null) {
-                LogWriteTag.Dispose ();
-                LogWriteTag = null;
-            }
-
-            if (UserIdTextField != null) {
-                UserIdTextField.Dispose ();
-                UserIdTextField = null;
-            }
-        }
-    }
+			if (UserIdTextField != null) {
+				UserIdTextField.Dispose ();
+				UserIdTextField = null;
+			}
+		}
+	}
 }