+++ SDK/AppCenter/Microsoft.AppCenter.Windows.Shared/Storage/IStorage.cs
@@ -74,5 +74,19 @@ namespace Microsoft.AppCenter.Storage
         /// <param name="timeout">The maximum amount of time to wait for remaining tasks</param>
         /// <returns>True if remaining tasks completed in time; false otherwise</returns>
         Task<bool> ShutdownAsync(TimeSpan timeout);
+
+        /// <summary>
+        /// Set the maximum size of the storage.
+        /// </summary>
+        /// <remarks>
+        /// This only sets the maximum size of the database, but App Center modules might store additional data.
+        /// The value passed to this method is not persisted on disk. The default maximum database size is 10485760 bytes (10 MiB).
+        /// </remarks>
+        /// <param name="sizeInBytes">
+        /// Maximum size of the storage in bytes. This will be rounded up to the nearest multiple of a SQLite page size (default is 4096 bytes).
+        /// Values below 20,480 bytes (20 KiB) will be ignored.
+        /// </param>
+        /// <returns><code>true</code> if changing the size was successful.</returns>
+        Task<bool> SetMaxStorageSizeAsync(long sizeInBytes);
     }
 }