+++ SDK/AppCenterDistribute/Microsoft.AppCenter.Distribute.Shared/DistributeDelegates.cs
@@ -7,4 +7,14 @@ namespace Microsoft.AppCenter.Distribute
     /// Release available callback.
     /// </summary>
     public delegate bool ReleaseAvailableCallback(ReleaseDetails releaseDetails);
+
+    /// <summary>
+    /// App will close callback.
+    /// </summary>
+    public delegate void WillExitAppCallback();
+
+    /// <summary>
+    /// No release available callback.
+    /// </summary>
+    public delegate void NoReleaseAvailableCallback();
 }