+++ SDK/AppCenterDistribute/Microsoft.AppCenter.Distribute.iOS/Distribute.cs
@@ -75,12 +75,42 @@ namespace Microsoft.AppCenter.Distribute
 
         static ReleaseAvailableCallback _releaseAvailableCallback;
 
+        static WillExitAppCallback _willExitAppCallback;
+        
+        static NoReleaseAvailableCallback _noReleaseAvailableCallback;
+
         static void SetReleaseAvailableCallback(ReleaseAvailableCallback releaseAvailableCallback)
         {
             lock (typeof(Distribute))
             {
                 _releaseAvailableCallback = releaseAvailableCallback;
                 if (_delegate == null && _releaseAvailableCallback != null)
+                {
+                    _delegate = new Delegate();
+                    iOSDistribute.SetDelegate(_delegate);
+                }
+            }
+        }
+
+        static void SetWillExitAppCallback(WillExitAppCallback willExitAppCallback)
+        {
+            lock (typeof(Distribute))
+            {
+                _willExitAppCallback = willExitAppCallback;
+                if (_delegate == null && _willExitAppCallback != null)
+                {
+                    _delegate = new Delegate();
+                    iOSDistribute.SetDelegate(_delegate);
+                }
+            }
+        }
+
+        static void SetNoReleaseAvailable(NoReleaseAvailableCallback noReleaseAvailable)
+        {
+            lock (typeof(Distribute))
+            {
+                _noReleaseAvailableCallback = noReleaseAvailable;
+                if (_delegate == null && _noReleaseAvailableCallback != null)
                 {
                     _delegate = new Delegate();
                     iOSDistribute.SetDelegate(_delegate);
@@ -152,6 +182,16 @@ namespace Microsoft.AppCenter.Distribute
                     return _releaseAvailableCallback(releaseDetails);
                 }
                 return false;
+            }
+
+            public override void WillExitApp(iOSDistribute distribute)
+            {
+                _willExitAppCallback?.Invoke();
+            }
+
+            public override void OnNoReleaseAvailable(MSACDistribute distribute)
+            {
+                _noReleaseAvailableCallback?.Invoke();
             }
         }
     }