+++ Tests/Microsoft.AppCenter.NET/Properties/AssemblyInfo.cs
@@ -36,8 +36,8 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("0.0.0.0")]
-[assembly: AssemblyFileVersion("4.0.0.0")]
-[assembly: AssemblyInformationalVersion("4.0.0-SNAPSHOT")]
+[assembly: AssemblyFileVersion("4.1.0.0")]
+[assembly: AssemblyInformationalVersion("4.1.0-SNAPSHOT")]
 
 [assembly: InternalsVisibleTo("Microsoft.AppCenter.Test.Windows")]
 [assembly: InternalsVisibleTo("Microsoft.AppCenter.Analytics.Test.Windows")]